Add files via upload
The script will now print the name of the mountain profile, when given such a name or will try to obtain the name form a spatial repository. The script will try to find the highest point on the profile and uses the tidygeocoder package to get its geographical name from OpenStreetMap.